refactor(#1704): campsite split — move joined_scan_stream.rs inline tests to a sibling file
Verbatim move (#1116/#1135). joined_scan_stream.rs was 793/800 lines, leaving no
room for the #1704 error seam; it is now 478.
